+++ lib/super_settings/storage/redis_storage.rb
@@ -23,19 +23,14 @@ module SuperSettings
     #
     # @example
     #   SuperSettings::Storage::RedisStorage.redis = ConnectionPool.new(size: 5) { Redis.new(url: ENV["REDIS_URL"]) }
-    class RedisStorage
-      include Storage
+    class RedisStorage < StorageAttributes
+      include Transaction
 
       SETTINGS_KEY = "SuperSettings.settings"
       UPDATED_KEY = "SuperSettings.order_by_updated_at"
 
-      class HistoryStorage
+      class HistoryStorage < HistoryAttributes
         HISTORY_KEY_PREFIX = "SuperSettings.history"
-
-        include SuperSettings::Attributes
-
-        attr_accessor :key, :value, :changed_by, :deleted
-        attr_reader :created_at
 
         class << self
           def find_all_by_key(key:, offset: 0, limit: nil)
@@ -55,10 +50,8 @@ module SuperSettings
             record
           end
 
-          def destroy_all_by_key(key)
-            RedisStorage.transaction do |redis|
-              redis.del("#{HISTORY_KEY_PREFIX}.#{key}")
-            end
+          def destroy_all_by_key(key, redis)
+            redis.del("#{HISTORY_KEY_PREFIX}.#{key}")
           end
 
           def redis_key(key)
@@ -66,24 +59,16 @@ module SuperSettings
           end
         end
 
-        def initialize(*)
-          @deleted = false
-          super
-        end
-
-        def created_at=(val)
-          @created_at = SuperSettings::Coerce.time(val)
-        end
-
         def save!
           raise ArgumentError.new("Missing key") if Coerce.blank?(key)
-          RedisStorage.transaction do |redis|
-            redis.lpush(self.class.redis_key(key), payload_json.to_json)
+
+          RedisStorage.transaction do |changes|
+            changes << self
           end
         end
 
-        def deleted?
-          !!@deleted
+        def save_to_redis(redis)
+          redis.lpush(self.class.redis_key(key), payload_json.to_json)
         end
 
         private
@@ -98,9 +83,6 @@ module SuperSettings
           payload
         end
       end
-
-      attr_reader :key, :raw_value, :description, :value_type, :updated_at, :created_at
-      attr_accessor :changed_by
 
       class << self
         def all
@@ -152,26 +134,15 @@ module SuperSettings
           end
         end
 
-        def transaction(&block)
-          if Thread.current[:super_settings_transaction_redis]
-            block.call(Thread.current[:super_settings_transaction_redis])
-          else
-            begin
-              with_redis do |redis|
-                redis.multi do |multi_redis|
-                  Thread.current[:super_settings_transaction_redis] = multi_redis
-                  Thread.current[:super_settings_transaction_after_commit] = []
-                  block.call(multi_redis)
-                end
-                after_commits = Thread.current[:super_settings_transaction_after_commit]
-                Thread.current[:super_settings_transaction_after_commit] = nil
-                after_commits.each(&:call)
+        def save_all(changes)
+          with_redis do |redis|
+            redis.multi do |multi_redis|
+              changes.each do |object|
+                object.save_to_redis(multi_redis)
               end
-            ensure
-              Thread.current[:super_settings_transaction_redis] = nil
-              Thread.current[:super_settings_transaction_after_commit] = nil
             end
           end
+          true
         end
 
         protected
@@ -185,7 +156,7 @@ module SuperSettings
         def load_from_json(json)
           attributes = JSON.parse(json)
           setting = new(attributes)
-          setting.send(:set_persisted!)
+          setting.persisted = true
           setting
         end
       end
@@ -206,74 +177,22 @@ module SuperSettings
         HistoryStorage.create!(key: key, value: value, deleted: deleted, changed_by: changed_by, created_at: created_at)
       end
 
-      def save!
-        self.updated_at ||= Time.now
-        self.created_at ||= updated_at
-        self.class.transaction do |redis|
-          redis.hset(SETTINGS_KEY, key, payload_json)
-          redis.zadd(UPDATED_KEY, updated_at.to_f, key)
-          set_persisted!
-        end
-        true
+      def save_to_redis(redis)
+        redis.hset(SETTINGS_KEY, key, payload_json)
+        redis.zadd(UPDATED_KEY, updated_at.to_f, key)
       end
 
       def destroy
-        self.class.transaction do |redis|
-          redis.hdel(SETTINGS_KEY, key)
-          redis.zrem(UPDATED_KEY, key)
-          HistoryStorage.destroy_all_by_key(key)
+        self.class.with_redis do |redis|
+          redis.multi do |multi_redis|
+            multi_redis.hdel(SETTINGS_KEY, key)
+            multi_redis.zrem(UPDATED_KEY, key)
+            HistoryStorage.destroy_all_by_key(key, multi_redis)
+          end
         end
-      end
-
-      def key=(value)
-        @key = (Coerce.blank?(value) ? nil : value.to_s)
-      end
-
-      def raw_value=(value)
-        @raw_value = (Coerce.blank?(value) ? nil : value.to_s)
-      end
-
-      def value_type=(value)
-        @value_type = (Coerce.blank?(value) ? nil : value.to_s)
-      end
-
-      def description=(value)
-        @description = (Coerce.blank?(value) ? nil : value.to_s)
-      end
-
-      def deleted=(value)
-        @deleted = Coerce.boolean(value)
-      end
-
-      def created_at=(value)
-        @created_at = SuperSettings::Coerce.time(value)
-      end
-
-      def updated_at=(value)
-        @updated_at = SuperSettings::Coerce.time(value)
-      end
-
-      def deleted?
-        !!@deleted
-      end
-
-      def persisted?
-        !!@persisted
       end
 
       private
-
-      def after_commit(&block)
-        if Thread.current[:super_settings_transaction_after_commit]
-          Thread.current[:super_settings_transaction_after_commit] << block
-        else
-          block.call
-        end
-      end
-
-      def set_persisted!
-        @persisted = true
-      end
 
       def payload_json
         payload = {