+++ spec/dummy/db/migrate/20240928235932_create_super_settings.super_settings.rb
@@ -1,13 +1,6 @@
 # frozen_string_literal: true
 
-# This migration comes from super_settings (originally 20210414004553)
-# Needed for the super_settings gem to maintain backward compatibility with Rails 4.2
-migration_class = ActiveRecord::Migration
-if migration_class.respond_to?(:[])
-  migration_class = migration_class[4.2]
-end
-
-class CreateSuperSettings < migration_class
+class CreateSuperSettings < ActiveRecord::Migration[5.0]
   def up
     create_table :super_settings do |t|
       t.string :key, null: false, limit: 190, index: {unique: true}