+++ spec/super_settings/application_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "../../spec_helper"
+require_relative "../spec_helper"
 
 describe SuperSettings::Application do
   it "should render the application HTML" do