+++ CHANGELOG.md
@@ -12,10 +12,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Isolated of CSS classes in the web UI to prevent conflicts with other CSS libraries.
 - Dark mode support in web UI.
 - Added ability to embed the web UI in a view to allow tighter integration with your application's UI.
+- Added storage adapter for storing settings in an S3 object.
+- Added abstract storage adapter for storing settings in a JSON file.
 
 ### Fixed
 
 - Changing a key now works as expected. Previously, a new setting was created with the new key and the old setting was left unchanged. Now, the old setting is properly marked as deleted.
+
+### Removed
+
+- Rails 4.2 support has been removed.
 
 ## 1.0.2